move adminCmds
